% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{attachStateDependencies}
\alias{attachStateDependencies}
\title{Searches a HTML fragment for dependencies.}
\usage{
attachStateDependencies(...)
}
\value{
A HTML tagList.
}
\description{
Searches a HTML fragment for dependencies.
}
\keyword{html}
\keyword{internal}
