% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{cns}
\alias{cns}
\title{Generates a namespaced id placeholder to allow traditional shiny widgets
  to be used as part of web component templates. Allows those widgets
  to be accessed in the shiny server function via input${componentID}_{id}}
\usage{
cns(id)
}
\arguments{
\item{id}{The inputId of the shiny widget.}
}
\value{
A namespaced Id
}
\description{
Generates a namespaced id placeholder to allow traditional shiny widgets
  to be used as part of web component templates. Allows those widgets
  to be accessed in the shiny server function via input${componentID}_{id}
}
