% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{componentScript}
\alias{componentScript}
\title{Runs Js code in the component scope}
\usage{
componentScript(session = getDefaultReactiveDomain(), inputId, callback)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.
Default is getDefaultReactiveDomain().}

\item{inputId}{The type of the component that was used in registerWebComponent
to register the component.}

\item{callback}{Named attributes to update and corresponding values.}
}
\description{
Runs Js code in the component scope
}
