% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropNulls}
\alias{dropNulls}
\title{Remove NULL entries.}
\usage{
dropNulls(x)
}
\arguments{
\item{x}{The object to process. Non list or vector objects are returned
with no changes.}
}
\value{
The give object x with all NULL entries removed.
}
\description{
Removes null values from a given list or vector. If the given object is of
  a diferent type, the same object is returned with no changes.
}
\keyword{internal}
\keyword{utils}
