% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{getComponent}
\alias{getComponent}
\title{Get a registered component.}
\usage{
getComponent(componentClass)
}
\arguments{
\item{componentClass}{The componentClass of a registered component.}
}
\value{
A component object.
}
\description{
Returns a object form of a registered component by its componentClass.
}
\keyword{components}
\keyword{internal}
