% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{htmlClassName}
\alias{htmlClassName}
\title{Generate a valid HTML class name.}
\usage{
htmlClassName(string)
}
\arguments{
\item{string}{The string to parse into a valid class name.}
}
\value{
A valid HTML class name (letters only, no spaces).
}
\description{
Converts a string into a valid format that can be used as a class name.
}
\keyword{html}
\keyword{internal}
