% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{htmlComponentTag}
\alias{htmlComponentTag}
\title{Generate a custom HTML tag.}
\usage{
htmlComponentTag(htmlTagName, inputId, ...)
}
\arguments{
\item{htmlTagName}{A valid HTML tag name to use in the template}

\item{inputId}{The tag ID attribute}

\item{...}{Attributes and content of the web component HTML wrapper tag}
}
\value{
A HTML tagList.
}
\description{
Generates a HTML custom tag with a defined ID, content and attributes.
Used in scaffoldWC().
}
\keyword{html}
\keyword{internal}
