% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{htmlWCTagName}
\alias{htmlWCTagName}
\title{Generate a valid HTML tag name.}
\usage{
htmlWCTagName(string)
}
\arguments{
\item{string}{The string to parse into a valid HTML tag name}
}
\value{
A valid HTML tag name.
}
\description{
Converts a string into a valid format that can be used as a HTML tag name
}
\keyword{html}
\keyword{internal}
