% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.ShinySession}
\alias{is.ShinySession}
\title{Tests whether the object is a valid ShinySession object.}
\usage{
is.ShinySession(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
TRUE or FALSE depending if the object is a valid ShinySession.
}
\description{
Tests whether the object is a valid ShinySession object.
}
\keyword{internal}
\keyword{utils}
