% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{listComponents}
\alias{listComponents}
\title{List registered component templates}
\usage{
listComponents()
}
\value{
A named list of css templates and specific values.
}
\description{
Lists all available component templates.
}
\keyword{breakpoint_system}
\keyword{breakpoints}
