% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listRenderTags}
\alias{listRenderTags}
\title{Get renderable tags from a object.}
\usage{
listRenderTags(object)
}
\arguments{
\item{object}{A object containing multiple entries. Tipically a HTML object.}
}
\value{
A HTML tagList.
}
\description{
Extracts and returns all HTML tags of an object that can be rendered.
}
\keyword{internal}
\keyword{utils}
