% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{registerComponent}
\alias{registerComponent}
\title{Adds the given element into the components list}
\usage{
registerComponent(componentClass, innerHTML)
}
\arguments{
\item{componentClass}{A unique identifier for the component template. Used later to
generate instances of the component. Acts similar to a class identifier.}

\item{innerHTML}{The HTML structure of the component}
}
\description{
Adds the given element into the components list
}
