% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{replacePlaceholders}
\alias{replacePlaceholders}
\title{Converts bind argument names to html data attributes.}
\usage{
replacePlaceholders(html, id)
}
\value{
A HTML tagList.
}
\description{
Converts bind argument names to html data attributes.
}
\keyword{html}
\keyword{internal}
