% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{scaffoldWC}
\alias{scaffoldWC}
\title{Generates a HTML tag with the required web component creation code,
  global dependencies, shiny bindings and tag content.}
\usage{
scaffoldWC(inputId, innerHTML, webClassName, htmlTagName, initialState, ...)
}
\arguments{
\item{inputId}{Component Id}

\item{innerHTML}{A valid HTML fragment to use as the component shadow DOM}

\item{webClassName}{A valid web class name to use in the template}

\item{htmlTagName}{A valid HTML tag name to use in the template}

\item{initialState}{Initial state of the component binds}

\item{...}{Attributes and content of the web component HTML wrapper tag}
}
\value{
A valid HTML dependency to add to a UI object
}
\description{
Generates a HTML tag with the required web component creation code,
  global dependencies, shiny bindings and tag content.
}
