% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{slot}
\alias{slot}
\title{HTML slot tag.}
\usage{
slot(name)
}
\arguments{
\item{name}{The name of the slot that can recieve future content.}
}
\value{
a HTML tag that can be added to the page.
}
\description{
A slot HTML tag with a given slot name. Can be used to define template areas
where content will be added in the future.
}
