% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{slotIn}
\alias{slotIn}
\title{Slot into component slot.}
\usage{
slotIn(name, ...)
}
\arguments{
\item{name}{The web component slot where the element should be added.}

\item{...}{The content and attributes to be slotIn into the web component.}
}
\value{
a HTML tag that can be added to the page.
}
\description{
Wraps a given HTML tag into a valid HTML element what can be slotIn into a
  specific slot in a web component.
}
\keyword{internal}
\keyword{namespaces}
\keyword{utils}
