% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{slotted}
\alias{slotted}
\title{Wraps a given HTML tag into a valid HTML element what can be slotted into a
  specific slot in a web component.}
\usage{
slotted(name, ...)
}
\arguments{
\item{name}{The web component slot where the element should be added.}

\item{...}{The content and attributes to be slotted into the web component.}
}
\value{
a HTML tag that can be added to the page.
}
\description{
Wraps a given HTML tag into a valid HTML element what can be slotted into a
  specific slot in a web component.
}
