% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{updateComponent}
\alias{updateComponent}
\title{Update a component state.}
\usage{
updateComponent(session = getDefaultReactiveDomain(), inputId, ...)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.
Default is getDefaultReactiveDomain().}

\item{inputId}{The id of the `component` object.}

\item{...}{Named attributes to update and corresponding values.}
}
\value{
No return value, called for side effects.
}
\description{
Update the component state with the corresponding inputId.
}
\keyword{component}
\keyword{components}
