% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{useComponent}
\alias{useComponent}
\title{Allows using a registered component in any place where a shiny UI can be
  used. Usage is similar to base shiny UI input widgets.}
\usage{
useComponent(inputId, componentClass, defaultState, ...)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{componentClass}{The class of the component that was used in
registerWebComponent to register the component.}

\item{defaultState}{Initial state for the component}

\item{...}{Aditional attributes to be added to the component HTML tag or/and
slotted components}
}
\value{
a HTML tag that can be added to the page.
}
\description{
Allows using a registered component in any place where a shiny UI can be
  used. Usage is similar to base shiny UI input widgets.
}
