% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{useShinyBound}
\alias{useShinyBound}
\title{Generate a shiny bound html dependency.}
\usage{
useShinyBound()
}
\value{
A HTML dependency object that can be added to a UI defnition.
}
\description{
Generates a html dependency to load and track shinyBound web dependencies.
}
\keyword{dependencies}
\keyword{internal}
\keyword{templates}
