% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateSessionObject}
\alias{validateSessionObject}
\title{Is session a ShinySession object.}
\usage{
validateSessionObject(
  session,
  label = as.character(sys.call(sys.parent())[[1]])
)
}
\arguments{
\item{session}{Any session object.}

\item{label}{The caller function's name}
}
\value{
No return value, called for side effects.
}
\description{
Check that an object is a ShinySession object, and give an informative
}
\keyword{internal}
\keyword{utils}
