% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{webComponentBindings}
\alias{webComponentBindings}
\title{Generates a HTML fragment with the required web component creation code.
  Used in scaffoldWC().}
\usage{
webComponentBindings(
  template,
  webClassName,
  htmlTagName,
  innerHTML,
  initialState
)
}
\arguments{
\item{template}{Path to an HTML template file}

\item{webClassName}{A valid web class name to use in the template}

\item{htmlTagName}{A valid HTML tag name to use in the template}

\item{innerHTML}{A valid HTML fragment to use as the component shadow DOM}

\item{initialState}{Initial state of the component binds}
}
\value{
A valid HTML dependency to add to a UI object
}
\description{
Generates a HTML fragment with the required web component creation code.
  Used in scaffoldWC().
}
