% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{componentScript}
\alias{componentScript}
\title{Run JS code under a component scope.}
\usage{
componentScript(session = getDefaultReactiveDomain(), inputId, callback)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.
Default is getDefaultReactiveDomain().}

\item{inputId}{The type of the component that was used in registerComponent
to register the component.}

\item{callback}{Named attributes to update and corresponding values.}
}
\value{
No return value, called for side effects.
}
\description{
Run JS code under the component scope. The provided callback will have
  access to the `this` JavaScript property as if running in the
  component root.
}
\keyword{component}
\keyword{components}
