% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{shinyBindings}
\alias{shinyBindings}
\title{Generate a Shiny Bind.}
\usage{
shinyBindings(template, webClassName, htmlTagName)
}
\arguments{
\item{template}{Path to an HTML template file}

\item{webClassName}{A valid web class name to use in the template}

\item{htmlTagName}{A valid HTML tag name to use in the template}
}
\value{
A HTML tagList.
}
\description{
Generates a HTML fragment that can be added to a page to initialize a
specific set of shiny bindings for a component. Used in scaffoldWC().
}
