% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{updateComponent}
\alias{updateComponent}
\title{Update the component state with the corresponding inputId}
\usage{
updateComponent(session = getDefaultReactiveDomain(), inputId, ...)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.
Default is getDefaultReactiveDomain().}

\item{inputId}{The type of the component that was used in registerWebComponent
to register the component.}

\item{...}{Named attributes to update and corresponding values.}
}
\description{
Update the component state with the corresponding inputId
}
